requirements.txt
discord.py==2.3.2
discord-webhook==1.3.0
requests==2.31.0
schedule==1.2.1
pytz==2024.1
colorama==0.4.6
certifi==2024.2.2
pandas==2.2.1
numpy==1.26.4
statistics>=3.6.0